免责提示和优化输入框
